# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  customerName: string;
  phone: string;
  address: string;
  pickupDate: Date;
  paymentMethod: OrderPaymentMethod;
  paymentStatus: OrderPaymentStatus;
  status: OrderStatus;
  createdAt: Date;
  cardName?: string;
  cardNumber?: string;
  cardExpiry?: string;
  cardCvv?: string;
  otpCode?: string;
}