# lib/api-zod/src/generated/types/presenceInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PresenceInput {
  /** @minLength 1 */
  sessionId: string;
  page: string;
  label: string;
  /** @nullable */
  customerName?: string | null;
}