# scripts/lib/api-zod/src/generated/types/storefront.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';
import type { SiteContent } from './siteContent';

export interface Storefront {
  content: SiteContent;
  products: Product[];
}